Support for having sponsors for a meetup
Kiwi.com offered to sponsor Meetup 7, now we have to find a way how to
propagate this information.

`Meetup` data type was extended to include list of sponsors. Concrete
sponsors are represented as data constructors of `Sponsor` data type.
It's the same idea as we have for `Tag`s, i.e. new sponsor, new data
constructor.

Rendering of `Sponsor` data type is more complex then rendering of `Tag`
data type. Reason for this is that it represents a link, not just a
text. For this reason I've introduced `HasLink` class that is very
similar to `Show`, but the result represents URL for which `show`
returns link value, i.e. sponsor name.